Tweaks to build scripts and VSCode tasks

+++ dev/make-package.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 [ "$1" == "" ] && { echo Missing package name; exit 1; }
+[ "$2" != "" ] && { echo Only one package name allowed; exit 1; }
 
 rm -rf {build,install}/"$1"
 source ./dev/setup.bash
